[WIP] Use twitter REST api (search api) for counts
Twitter count-json isn't working anymore. It's done, finished. So now we have to do the counts ourselves with an iterative search on URL via the Twitter REST API. How awkward....and so very very very inefficient....

It works, but you'll need to set up an app and use the consumer key/secret + access token/secret. Be mindfull of the API limits.

Needs some more testing. A variant of this branch is used @ http://mediaweb.nl.